+++ docs/changelog.rst
@@ -11,6 +11,10 @@ Unreleased
   :func:`gidgethub.sansio.format_url`
   (`PR #234 <https://github.com/gidgethub/gidgethub/pull/234>`_)
 
+- Use httpx2 as a drop-in replacement for httpx backend.
+  The "httpx" extra was updated to use httpx2 and a new httpx2 extra
+  was also added.
+
 5.4.0
 -----